requirements.txt
deepdiff>=7.0.1
ldap3>=2.9.1
pymysql>=1.1.1
python-dotenv>=1.0.1
ratelimit>=2.2.1
Requests>=2.32.3
